refactor: rename Setup namespace to Settings for consistency
