+++ src/Ivy.Tendril.TeamIvyConfig/config.yaml
@@ -1,4 +1,4 @@
-codingAgent: opencode
+codingAgent: copilot
 jobTimeout: 30
 staleOutputTimeout: 10
 gitTimeout: 10
