+++ src/Ivy.Tendril/Apps/Settings/AdvancedSetupView.cs
@@ -1,7 +1,7 @@
 using Ivy.Tendril.Services;
 using Ivy.Tendril.Helpers;
 
-namespace Ivy.Tendril.Apps.Setup;
+namespace Ivy.Tendril.Apps.Settings;
 
 public class AdvancedSetupView : ViewBase
 {
